moved Stockfish files into an 'Engine' package
