+++ src/main/java/Controller/Engine/StockFishOutput.java
@@ -1,4 +1,4 @@
-package Controller;
+package Controller.Engine;
 
 import Model.Pieces.PieceType;
 import Model.Position;
