+++ src/main/java/Controller/Engine/Stockfish.java
@@ -1,4 +1,4 @@
-package Controller;
+package Controller.Engine;
 
 import Model.Game;
 import Model.Pieces.PieceType;
